move: reader 폴더로 분리
