+++ src/main/java/codesquad/handler/RedirectStaticFileHandler.java
@@ -1,6 +1,7 @@
 package codesquad.handler;
 
 import codesquad.http.*;
+import codesquad.reader.StaticFileReaderSpec;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
