+++ src/main/java/codesquad/reader/StaticFileReader.java
@@ -1,4 +1,4 @@
-package codesquad.handler;
+package codesquad.reader;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
